+++ src/pages/Favorites/index.js
@@ -189,7 +189,7 @@ class Favorites extends React.Component {
 const mapStateToProps = (state) => {
     return {
         user: state.user,
-        favorProfs: state.favorProfs,
+        favorProfs: state.user.favorProfs,
     };
 };