+++ src/pages/Search/index.js
@@ -1,12 +1,20 @@
-import React from 'react'
-import { View, Text, Image, TextInput, TouchableOpacity, Keyboard, TouchableWithoutFeedback } from 'react-native';
+import React from "react";
+import {
+    View,
+    Text,
+    Image,
+    TextInput,
+    TouchableOpacity,
+    Keyboard,
+    TouchableWithoutFeedback,
+} from "react-native";
 import { LinearGradient } from "expo-linear-gradient";
-import { Feather } from '@expo/vector-icons';
-import { useNavigation } from '@react-navigation/native';
+import { Feather } from "@expo/vector-icons";
+import { useNavigation } from "@react-navigation/native";
 
-import globalStyles from '../../globalStyle/globalStyles';
-import styles from './styles';
-import search from '../../assets/magGlass.png';
+import globalStyles from "../../globalStyle/globalStyles";
+import styles from "./styles";
+import search from "../../assets/magGlass.png";
 
 export default function Search() {
     const navigation = useNavigation();
@@ -14,6 +22,11 @@ export default function Search() {
     function goBack() {
         navigation.goBack();
     }
+
+    function toCards() {
+        navigation.navigate("ProfCards");
+    }
+
     return (
         <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
             <View style={globalStyles.container}>
@@ -23,10 +36,9 @@ export default function Search() {
                     end={[1.0, 0.5]}
                     locations={[0.0, 1.0]}
                     style={styles.gradientBar}
-                >
-                </LinearGradient>
+                ></LinearGradient>
                 <TouchableOpacity style={styles.xIcon} onPress={goBack}>
-                    <Feather name='x' size={30} />
+                    <Feather name="x" size={30} />
                 </TouchableOpacity>
 
                 <View style={styles.searchContainer}>
@@ -35,17 +47,16 @@ export default function Search() {
                         placeholder="Busque palavras chave"
                         underlineColorAndroid="#333"
                         accessibilityLabel="Search Input"
-                        onChange={() => { }}
+                        onChange={() => {}}
                     />
                 </View>
 
-
                 <View style={styles.searchBtn}>
-                    <TouchableOpacity style={styles.btn}>
+                    <TouchableOpacity style={styles.btn} onPress={toCards}>
                         <Text style={styles.text}>Aplicar</Text>
                     </TouchableOpacity>
                 </View>
             </View>
         </TouchableWithoutFeedback>
-    )
+    );
 }