+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 -   Search page
 -   Header navigation
+-   Different Redux for all teachers and user 
+-   Context to user logged
 
 
 ## Release [0.0.7] - 22-04-2020